requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
bootstrap-admin==0.3.7.1
celery==3.1.24
click==6.6
colorama==0.3.7
demjson==2.2.4
Django==1.10.4
django-appconf==1.0.2
django-avatar==3.1.0
django-bootstrap==0.2.4
django-bootstrap3==8.0.0
django-celery==3.1.17
django-ckeditor==5.1.1
django-contrib-comments==1.7.3
django-crispy-forms==1.6.1
django-disqus==0.5
django-fluent-comments==1.3
django-model-utils==2.6
django-redis==4.5.0
django-tag-parser==3.1
django-threadedcomments==1.0.1
django-users2==0.2.1
DjangoUeditor==1.8.143
ecdsa==0.13
et-xmlfile==1.0.1
Fabric==1.12.0
Flask==0.11.1
funcsigs==1.0.2
image==1.5.5
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8
kombu==3.0.37
Markdown==2.6.7
MarkupSafe==0.23
mock==2.0.0
MySQL-python==1.2.3
olefile==0.43
openpyxl==2.4.0
paramiko==1.17.2
pbr==1.10.0
Pillow==4.0.0
pycrypto==2.6.1
python-akismet==0.2.3
pytz==2016.7
qrcode==5.0
redis==2.10.5
requests==2.12.4
shadowsocks==2.8.2
six==1.10.0
South==1.0.2
stevedore==1.18.0
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.11.11
